Version 0.1.0 - Project Start
I made a nice little plugin version of my Self Updatable E2 script & included it in here.

Now, this script is self-updatable, whenever I make new additions to it.

Starting at 0.1.0, the self-updatable feature is the only thing that is quite possibly functional. But, more features will be added in as time marches on.

Watch this space. =^/,..,^=