Added Automatic Parenting & Smart Entity Discovery
No need to manually parent the E2 chip to the locomotive's body. That is all taken care of, by the Automatic Parenting.

You no longer need to spawn in your bogies. The E2 script takes care of that, by way of Smart Entity Discovery.

Say "goodbye" to the purists what would otherwise yell at ya for having incorrect bogies on your locomotives.

I have added in an extra bonus, where all you need to do is spawn in your locomotive body blank of choice (currently, only supports the SD40), attach this E2 script to the body (preferably in an inconspicuous place) & this E2 script will automatically spawn the bogies in the appropriate places, with the appropriate constraints.